requirements.txt
asgiref==3.4.1
Django==4.0
django-crispy-forms==1.13.0
psycopg2==2.9.3
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5